Added models, colliders, animations, and positions
Added generic models for the players and added it to the player prefab. Also added colliders and rigid body for physics to allow the players to stand on the plane. Batter spawns in a bat, however currently not rotated correctly. Batter and pitcher have animations triggered by right click currently. Added the main animator named Animator to handle all animations and triggers.