Added pitcher prefab and animation transitions
Added pitcher prefab and animations, transitions, and triggers to make the pitcher throw the ball and the batter swing. I have to mess with the ball rigidbody because currenly it just falls out of the pitchers hand. The players also still do not have the proper orientation and are facing different directions.