Added animation events, ball script, and updated animations
Added animation events so the ball will in theory leave the pitchers hand at the correct time. I added a script to the baseball to help with the physics. However, right now the ball just sort of floats out of the players hand, I can't get it to stay no matter what I have tried so far